Unnecessary optimization but I couldn't help myself.
Updated the README for the exercise
.

Added an example.

Added the exercise to the config.json file.

Updated problem-specifications to 1.0.1 and added a test against a DNA nucleotide. Also added a space after the `if`s in the code. Lastly, changed the test iterable to contain null when the result should be an Error.